feat: Support shortcuts to enter the full screen mode
