+++ packages/full-screen/src/EnterFullScreen.tsx
@@ -7,11 +7,10 @@
  */
 
 import * as React from 'react';
-import { isMac } from '@react-pdf-viewer/core';
 import type { Store } from '@react-pdf-viewer/core';
 
 import { EnterFullScreenButton } from './EnterFullScreenButton';
-import { addFullScreenChangeListener, exitFullScreen, getFullScreenElement, requestFullScreen } from './fullScreen';
+import { useEnterFullScreen } from './useEnterFullScreen';
 import type { StoreProps } from './types/StoreProps';
 import type { Zoom } from './types/Zoom';
 
@@ -32,54 +31,7 @@ export const EnterFullScreen: React.FC<{
     onEnterFullScreen(zoom: Zoom): void;
     onExitFullScreen(zoom: Zoom): void;
 }> = ({ children, enableShortcuts, store, onEnterFullScreen, onExitFullScreen }) => {
-    const pagesRef = React.useRef<HTMLElement | null>(
-        store.get('getPagesContainer') ? store.get('getPagesContainer')() : null
-    );
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const closeOtherFullScreen = (): Promise<any> => {
-        const pagesEle = pagesRef.current;
-        if (!pagesEle) {
-            return Promise.resolve();
-        }
-
-        const ele = getFullScreenElement();
-        return ele && ele !== pagesEle ? exitFullScreen(ele) : Promise.resolve();
-    };
-
-    const enterFullScreen = () => {
-        const pagesEle = pagesRef.current;
-        if (!pagesEle) {
-            return;
-        }
-
-        closeOtherFullScreen().then(() => {
-            requestFullScreen(pagesEle);
-        });
-    };
-
-    const onFullScreenChange = (): void => {
-        const ele = getFullScreenElement();
-        const isFullScreen = ele === pagesRef.current;
-        store.update('isFullScreen', isFullScreen);
-
-        const zoom = store.get('zoom');
-        if (zoom) {
-            isFullScreen ? onEnterFullScreen(zoom) : onExitFullScreen(zoom);
-        }
-    };
-
-    const handlePagesContainer = (getPagesContainer: () => HTMLElement) => {
-        pagesRef.current = getPagesContainer();
-        addFullScreenChangeListener(onFullScreenChange);
-    };
-
-    React.useEffect(() => {
-        store.subscribe('getPagesContainer', handlePagesContainer);
-        return (): void => {
-            store.unsubscribe('getPagesContainer', handlePagesContainer);
-        };
-    }, []);
+    const { enterFullScreen } = useEnterFullScreen(store, onEnterFullScreen, onExitFullScreen);
 
     const defaultChildren = (props: RenderEnterFullScreenProps) => (
         <EnterFullScreenButton enableShortcuts={enableShortcuts} onClick={props.onClick} />
